requirements.txt
alembic==1.7.4
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.2.0
redis==3.5.3
rq==1.10.0
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
WTForms==2.3.3
